+++ src/index.test.ts
@@ -34,6 +34,7 @@ import {
   getSourceUrl,
   getSourceIri,
   saveSolidDatasetAt,
+  createEmptyContainerAt,
   saveSolidDatasetInContainer,
   saveAclFor,
   deleteAclFor,
@@ -201,6 +202,7 @@ it("exports the public API from the entry file", () => {
   expect(getSourceUrl).toBeDefined();
   expect(getSourceIri).toBeDefined();
   expect(saveSolidDatasetAt).toBeDefined();
+  expect(createEmptyContainerAt).toBeDefined();
   expect(saveSolidDatasetInContainer).toBeDefined();
   expect(saveAclFor).toBeDefined();
   expect(deleteAclFor).toBeDefined();